Adicionando implementação de cesta de compras

+++ src/main/resources/scripts/migrations/V1.01__create_products_table.sql
@@ -1,10 +1,11 @@
-CREATE  TABLE MarketApp.products(
-         id      uuid NOT NULL primary key,
-         sku varchar(255) NOT NULL UNIQUE,
-         name varchar(255) NOT NULL,
-         price double not null,
-         amount bigint not null,
-         createdAt timestamp not null,
-         updatedAt timestamp not null
+CREATE TABLE MarketApp.products
+(
+    id        uuid           NOT NULL primary key,
+    sku       varchar(255)   NOT NULL UNIQUE,
+    name      varchar(255)   NOT NULL,
+    price     numeric(38, 2) not null,
+    amount    bigint         not null,
+    createdAt timestamp      not null,
+    updatedAt timestamp      not null
 );
